Add checking for whether the attachment is an image.
Added a templatetag to render an image tag for the attachment,
and a postsave hook for figuring out whether the attachment is an
image. If you're using a local filestore, the logic in the hook may
seem bizarre; it was written to support using an S3-backed filestore
for attachments.

Also added a basic template for rendering the upload form and all
attachments with one {% include %} tag.